functions/data_validation_function/requirements.txt
google-cloud-bigquery>=3.11.0
functions-framework>=3.10.0

